tests/wpt/web-platform-tests/css/css-text/parsing/white-space-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing white-space with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-white-space">
<meta name="assert" content="white-space supports only the grammar 'normal | pre | nowrap | pre-wrap | break-spaces | pre-line'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("white-space", "auto");
test_invalid_value("white-space", "normal pre");
test_invalid_value("white-space", "nowrap pre-wrap");
test_invalid_value("white-space", "pre-line break-spaces");
</script>
</body>
</html>
